0520 첨부파일(사진안나오고 파일명이 나옴)
