+++ caffeineD/src/main/java/co/dev/web/FrontController.java
@@ -15,6 +15,8 @@ import co.dev.web.board.BoardContentsControl;
 import co.dev.web.board.BoardDeleteControl;
 import co.dev.web.board.BoardInsertControl;
 import co.dev.web.board.BoardLoadControl;
+import co.dev.myPage.myPageLoadController;
+import co.dev.myPage.updateInfoControl;
 import co.dev.web.admin.userListControl;
 import co.dev.web.bookmark.BookmarkCheckControl;
 import co.dev.web.bookmark.BookmarkSelectControl;
