+++ caffeineD/src/main/java/co/dev/web/board/BoardDeleteControl.java
@@ -5,32 +5,23 @@ import java.io.IOException;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
 
 import co.dev.service.BoardService;
-import co.dev.vo.BoardVO;
-import co.dev.vo.UserVO;
 import co.dev.web.Controller;
 
-public class BoardDeleteControl implements Controller {	
+public class BoardDeleteControl implements Controller {
 	@Override
 	public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		HttpSession session = request.getSession();
-		UserVO vo = (UserVO) session.getAttribute("user");
-		String user = request.getParameter("user");
-		if(user.isEmpty()) {
-			request.setAttribute("error", "작성자를 입력하세요");
-			request.getRequestDispatcher("BoardDelete.jsp").forward(request, response);
-		}
-
-		BoardVO board = new BoardVO();
-//		board.setUser(user);
 		
-		BoardService service = new BoardService();
-		service.boardDelete(user);
+		String board_no = request.getParameter("board_no");
+		
+		
+		
 
-		request.setAttribute("user", user);
-		request.getRequestDispatcher("result/deleteBoardOutput.jsp").forward(request, response);
+
+		//response.sendRedirect("boardContents.do");
+//		request.setAttribute("user", user);
+		//request.getRequestDispatcher("result/deleteBoardOutput.jsp").forward(request, response);
 	}
 
 }
