+++ caffeineD/src/main/webapp/boardContents.jsp
@@ -34,12 +34,18 @@
 				</tbody>
 			</table>
 			<br>
+			<td>${contents.img }</td><br><br>
 			<td>${contents.content }</td>
 		</c:otherwise>
-	</c:choose><br><br>
-		<div style="margin-left:10%;">
-		<button onclick="location.href='updateBoard.jsp'" type="button">글수정</button> <button onclick="location.href='deleteBoard.jsp'" type="button">삭제</button>
-		</div>
+	</c:choose>
+	<br>
+	<br>
+	<div style="margin-left: 10%;">
+		<button onclick="location.href='updateBoard.jsp'" type="button">글수정</button>
+		<button onclick="location.href='deleteBoard.jsp'" type="button">삭제</button>
+				
+
+	</div>
 	<c:choose>
 		<c:when test="${empty comments }">
 			<h3>댓글이 없습니다.</h3>
@@ -51,7 +57,7 @@
 						<tr>
 							<th>번호</th>
 							<th>작성자</th>
-							<th>내용</th>							
+							<th>내용</th>
 					<tbody>
 						<c:forEach items="${comments }" var="comment">
 							<div style="margin-top: 100px;">
@@ -61,11 +67,8 @@
 									<td>${comment.comment_content }</td>
 								</tr>
 							</div>
-
-
 						</c:forEach>
 					</tbody>
-
 				</table>
 			</form>
 		</c:otherwise>
