+++ caffeineD/src/main/webapp/myComment.jsp
@@ -19,6 +19,6 @@
 		</table>
 	</form>
 	<button style="margin-top: 10px;" type="button"
-		onclick="location.href='myPage.jsp'">돌아가기</button>
+		onclick="location.href='myPage.do'">돌아가기</button>
 </body>
 </html>
