+++ caffeineD/src/main/webapp/updateInfo.jsp
@@ -10,11 +10,12 @@
 </head>
 <body>	
 	<h2>정보 수정</h2>
-	<form action="${pageContext.servletContext.contextPath }/updateInfo.do" method="post">
+	<form action="${pageContext.servletContext.contextPath }/updateInfo.do" method="post" enctype="Multipart/form-data">
 		아이디: <input type="text" name="id" readonly value="${user.id }"><br>
 		비밀번호: <input type="password" name="pwd" value="${user.pwd }"><br>
 		닉네임: <input type="text" name="nickname" value="${user.nickname }"><br>
 		연락처: <input type="text" name="tel" value="${user.tel }"><br>
+		프로필: <input type="file" name="img" value=img><br>
 		<input style="margin-top:10px;" type="submit" value="수정"> <button style="margin-top:10px;" type="button" onclick="location.href='myPage.jsp'">돌아가기</button>
 	</form>	
 </body>
